+++ stormify/src/jvmBasedMain/kotlin/onl/ycode/stormify/ConnectToNative.jvm.kt
@@ -138,7 +138,7 @@ internal actual fun <T : Any> tryReflection(type: KClass<T>): EntityMeta<T>? {
 
         val dbField = annotations.filterIsInstance<DbField>().firstOrNull()
         val jpaId = annotations.any { it.annotationClass.qualifiedName == "javax.persistence.Id" }
-        val jpaColumn = annotations.find { it.annotationClass.qualifiedName == "javax.persistence.Facet" }
+        val jpaColumn = annotations.find { it.annotationClass.qualifiedName == "javax.persistence.Column" }
         val jpaJoinColumn = annotations.find { it.annotationClass.qualifiedName == "javax.persistence.JoinColumn" }
         val jpaSequence = annotations.find { it.annotationClass.qualifiedName == "javax.persistence.SequenceGenerator" }